requirements.txt
fastapi
uvicorn
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-prometheus
prometheus-client
python-dotenv
sqlalchemy
mlflow